backend/requirements.txt
Django==2.2.15
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.2
pytz==2020.4
sqlparse==0.4.1
